fix: getIdBySlug in api slug2id
